backend/requirements/development.txt
-r ./base.txt

pytest-cov==3.0.0
pytest==6.2.5
pytest-django==4.4.0
pytest-mock==3.6.1

Faker==9.8.0
ipdb==0.13.9
factory_boy==3.2.1
freezegun==1.1.0
responses==0.15.0
